Second Efforts -- Creating Bios View -- In progress

+++ app/src/main/java/com/cyberfanta/torre_co_challenge_app/models/bio/EducationItem.java
@@ -11,7 +11,7 @@ public class EducationItem{
 	private boolean highlighted;
 
 	@SerializedName("weight")
-	private int weight;
+	private float weight;
 
 	@SerializedName("media")
 	private List<Object> media;
@@ -60,11 +60,11 @@ public class EducationItem{
 		return highlighted;
 	}
 
-	public void setWeight(int weight){
+	public void setWeight(float weight){
 		this.weight = weight;
 	}
 
-	public int getWeight(){
+	public float getWeight(){
 		return weight;
 	}
