+++ app/src/main/java/com/cyberfanta/torre_co_challenge_app/models/bio/JobsItem.java
@@ -17,7 +17,7 @@ public class JobsItem{
 	private String toMonth;
 
 	@SerializedName("weight")
-	private int weight;
+	private float weight;
 
 	@SerializedName("media")
 	private List<Object> media;
@@ -82,11 +82,11 @@ public class JobsItem{
 		return toMonth;
 	}
 
-	public void setWeight(int weight){
+	public void setWeight(float weight){
 		this.weight = weight;
 	}
 
-	public int getWeight(){
+	public float getWeight(){
 		return weight;
 	}
